% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_setup.R
\name{api_setup}
\alias{api_setup}
\title{setup the api}
\usage{
api_setup(api_key, region)
}
\arguments{
\item{api_key}{a character}

\item{region}{a character, must be one of "br", "eune", "euw", "kr", "lan", "las", "na", "oce", "ru" or "tr"}
}
\description{
Use this function to set up the usual region and key you want to use.
}
\details{
Get your key at \url{https://developer.riotgames.com/}
}

