% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champion.R
\name{champion}
\alias{champion}
\alias{champion_static}
\title{champion data}
\usage{
champion(id = NULL, freeToPlay = NULL)

champion_static(id = NULL, locale = NULL, version = NULL,
  dataById = NULL, champData = NULL)
}
\description{
Wraps the \href{https://developer.riotgames.com/api/methods#!/1077}{champion} and \href{https://developer.riotgames.com/api/methods#!/1055/3633}{static champion} api method
}

