% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champion_mastery.R
\name{champion_mastery}
\alias{champion_mastery}
\title{champion mastery data}
\usage{
champion_mastery(player_id, champion_id = NULL, type = NULL)
}
\description{
Wraps the \href{https://developer.riotgames.com/api/methods#!/1071}{champion mastery} api method
}

