% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/champions.R
\name{Champions}
\alias{Champions}
\alias{GetChampion}
\alias{GetChampionList}
\title{Functions for champions}
\usage{
GetChampionList(region = myRegion, apiKey = myAPIKey)

GetChampion(championId, region = myRegion, apiKey = myAPIKey)
}
\arguments{
\item{region}{A character, the region where to retrieve the data (irrelevant here).}

\item{apiKey}{A character, the API key delivered by Riot Games.}
}
\value{
A \code{tbl} (see \code{\link[dplyr]{tbl_df}}) containing some information for each champion (see Description)/
}
\description{
Returns some information about each champion :
\itemize{
  \item id : integer, assigned by Riot Games
  \item active: boolean,	indicates if the champion is active.
  \item botEnabled: boolean, bot enabled flag (for custom games).
  \item botMmEnabled: boolean, bot Match Made enabled flag (for Co-op vs. AI games).
  \item freeToPlay: boolean, indicates if the champion is free to play. Free to play champions are rotated periodically.
  \item rankedPlayEnabled: boolean, ranked play enabled flag.
}
}

