% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{game}
\alias{game}
\title{game data}
\usage{
game(summoner_id)
}
\arguments{
\item{summoner_id}{a numeric}
}
\description{
Wraps the \href{https://developer.riotgames.com/api/methods#!/1078}{team} api method
}

