% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{league}
\alias{league}
\title{league data}
\usage{
league(summoner_id = NULL, team_id = NULL, entry = NULL)
}
\arguments{
\item{summoner_id}{a vector of numerics}

\item{team_id}{a vector of numerics}

\item{entry}{a logical}
}
\description{
Wraps the \href{https://developer.riotgames.com/api/methods#!/985}{league} api method
}

