% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{static}
\alias{champion_static}
\alias{item_static}
\alias{language_strings_static}
\alias{languages_static}
\alias{map_static}
\alias{mastery_static}
\alias{realm_static}
\alias{rune_static}
\alias{static}
\alias{summoner_spell_static}
\alias{versions_static}
\title{static data}
\usage{
champion_static(id = NULL, locale = NULL, version = NULL,
  data_by_id = NULL, champ_data = NULL)

item_static(id = NULL, locale = NULL, version = NULL, item_data = NULL)

language_strings_static(locale = NULL, version = NULL)

languages_static()

map_static(locale = NULL, version = NULL)

mastery_static(id = NULL, locale = NULL, version = NULL,
  mastery_data = NULL)

realm_static()

rune_static(id = NULL, locale = NULL, version = NULL, rune_data = NULL)

summoner_spell_static(id = NULL, locale = NULL, version = NULL,
  spell_data = NULL)

versions_static()
}
\arguments{
\item{id}{a numeric, depends on the method call, can be a champion, item, mastery, rune or summoner spell id}

\item{locale}{a character}

\item{version}{a character}

\item{data_by_id}{a logical}

\item{champ_data}{a vector of characters}

\item{item_data}{a vector of characters}

\item{mastery_data}{a vector of characters}

\item{rune_data}{a vector of characters}

\item{spell_data}{a vector of characters}
}
\description{
Wraps the \href{https://developer.riotgames.com/api/methods#!/1055}{static-data} api method
}

