% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{stats data}
\usage{
stats(summoner_id, type = c("ranked", "summary"), season = NULL)
}
\arguments{
\item{summoner_id}{a numeric}

\item{type}{a character, either 'ranked' or 'summary'}

\item{season}{a character}
}
\description{
Wraps the \href{https://developer.riotgames.com/api/methods#!/1080}{stats} api method
}
\details{
\code{season} should be in the \code{SEASONYYYY} format, starting from 2014. Data is available for the
2013 season with \code{SEASON3}.
}

