% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summoner.R
\name{summoner}
\alias{summoner}
\title{summoner data}
\usage{
summoner(name = NULL, id = NULL, type = NULL)
}
\arguments{
\item{name}{a vector of names}

\item{id}{a vector of numerics}
}
\description{
Wraps the \href{https://developer.riotgames.com/api/methods#!/1079}{summoner} api method
}

