% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{team}
\alias{team}
\title{team data}
\usage{
team(summoner_id = NULL, team_id = NULL)
}
\arguments{
\item{summoner_id}{a vector of numerics}

\item{team_id}{a vector of numerics}
}
\description{
Wraps the \href{https://developer.riotgames.com/api/methods#!/986}{team} api method
}

